+++ build_daemon/lib/constants.dart
@@ -12,7 +12,9 @@ const optionsSkew = 'DIFFERENT OPTIONS';
 
 const buildModeFlag = 'build-mode';
 enum BuildMode {
+  // ignore: constant_identifier_names
   Manual,
+  // ignore: constant_identifier_names
   Auto,
 }