+++ build_daemon/lib/src/server.dart
@@ -6,6 +6,7 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
+import 'package:build_daemon/data/server_log.dart';
 import 'package:built_value/serializer.dart';
 import 'package:http_multi_server/http_multi_server.dart';
 import 'package:pool/pool.dart';
@@ -29,6 +30,8 @@ import 'managers/build_target_manager.dart';
 /// Handles notifying clients of logs and results for registered build targets.
 /// Note the server will only notify clients of pertinent events.
 class Server {
+  static final loggerName = 'BuildDaemonServer';
+
   final _isDoneCompleter = Completer();
   final BuildTargetManager _buildTargetManager;
   final _pool = Pool(1);
@@ -43,13 +46,20 @@ class Server {
 
   final _subs = <StreamSubscription>[];
 
-  Server(this._builder, Duration timeout, ChangeProvider changeProvider,
-      {Serializers? serializersOverride,
-      bool Function(BuildTarget, Iterable<WatchEvent>)? shouldBuild})
-      : _changeProvider = changeProvider,
+  final _outputStreamController = StreamController<ServerLog>();
+  late final Stream<ServerLog> _logs;
+
+  Server(
+    this._builder,
+    Duration timeout,
+    ChangeProvider changeProvider, {
+    Serializers? serializersOverride,
+    bool Function(BuildTarget, Iterable<WatchEvent>)? shouldBuild,
+  })  : _changeProvider = changeProvider,
         _serializers = serializersOverride ?? serializers,
         _buildTargetManager =
             BuildTargetManager(shouldBuildOverride: shouldBuild) {
+    _logs = _outputStreamController.stream;
     _forwardData();
 
     _handleChanges(changeProvider.changes);
@@ -71,8 +81,8 @@ class Server {
         dynamic request;
         try {
           request = _serializers.deserialize(jsonDecode(message as String));
-        } catch (_) {
-          print('Unable to parse message: $message');
+        } catch (e, s) {
+          _logMessage(Level.WARNING, 'Unable to parse message: $message', e, s);
           return;
         }
         if (request is BuildTargetRequest) {
@@ -90,7 +100,11 @@ class Server {
     });
 
     var server = _server = await HttpMultiServer.loopback(0);
-    serveRequests(server, handler);
+    // Serve requests in an error zone to prevent failures
+    // when running from another error zone.
+    runZonedGuarded(() => serveRequests(server, handler), (e, s) {
+      _logMessage(Level.WARNING, 'Error serving requests', e, s);
+    });
     return server.port;
   }
 
@@ -109,6 +123,7 @@ class Server {
     for (var sub in _subs) {
       await sub.cancel();
     }
+    await _outputStreamController.close();
     if (!_isDoneCompleter.isCompleted) _isDoneCompleter.complete();
   }
 
@@ -133,6 +148,12 @@ class Server {
         for (var channel in _interestedChannels) {
           channel.sink.add(message);
         }
+      }))
+      ..add(_logs.listen((log) {
+        var message = jsonEncode(_serializers.serialize(log));
+        for (var channel in _interestedChannels) {
+          channel.sink.add(message);
+        }
       }));
   }
 
@@ -153,4 +174,13 @@ class Server {
       await stop();
     }
   }
+
+  void _logMessage(Level level, String message,
+          [Object? error, StackTrace? stackTrace]) =>
+      _outputStreamController.add(ServerLog((b) => b
+        ..message = message
+        ..level = level
+        ..loggerName = loggerName
+        ..error = error?.toString()
+        ..stackTrace = stackTrace?.toString()));
 }