+++ build_runner/lib/src/daemon/change_providers.dart
@@ -3,9 +3,11 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:build_daemon/change_provider.dart';
+// ignore: implementation_imports
 import 'package:build_runner_core/src/asset_graph/graph.dart';
+// ignore: implementation_imports
 import 'package:build_runner_core/src/generate/build_definition.dart';
-import 'package:watcher/src/watch_event.dart';
+import 'package:watcher/watcher.dart' show WatchEvent;
 
 /// Continually updates the [changes] stream as watch events are seen on the
 /// input stream.