+++ build_runner/lib/src/server/asset_graph_handler.dart
@@ -7,7 +7,9 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:build/build.dart';
+// ignore: implementation_imports
 import 'package:build_runner_core/src/asset_graph/graph.dart';
+// ignore: implementation_imports
 import 'package:build_runner_core/src/asset_graph/node.dart';
 import 'package:glob/glob.dart';
 import 'package:shelf/shelf.dart' as shelf;