+++ build_runner/lib/src/server/server.dart
@@ -8,6 +8,7 @@ import 'dart:io';
 import 'package:build/build.dart';
 import 'package:build_runner/src/entrypoint/options.dart';
 import 'package:build_runner_core/build_runner_core.dart';
+// ignore: implementation_imports
 import 'package:build_runner_core/src/generate/performance_tracker.dart';
 import 'package:crypto/crypto.dart';
 import 'package:glob/glob.dart';
@@ -207,7 +208,7 @@ class BuildUpdatesWebSocketHandler {
 
   shelf.Handler createHandlerByRootDir(String rootDir) {
     if (!_internalHandlers.containsKey(rootDir)) {
-      var closureForRootDir = (WebSocketChannel webSocket, String protocol) =>
+      void closureForRootDir(WebSocketChannel webSocket, String protocol) =>
           _handleConnection(webSocket, protocol, rootDir);
       _internalHandlers[rootDir] = _handlerFactory(closureForRootDir,
           protocols: [_buildUpdatesProtocol]);