+++ build_daemon/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.0.2
+
+- Run `serveRequests` in an error zone and forward errors to the clients.
+
+## 3.0.1
+
+- Drop package:pedantic dependency and replace it with package:lints.
+
 ## 3.0.0
 
 - Migrate to null safety.