+++ build_runner_core/CHANGELOG.md
@@ -1,4 +1,10 @@
-## 7.2.1-dev
+## 7.2.2
+
+- Fix a bug compilation bug on the 2.14 sdk.
+
+## 7.2.1
+
+- Drop package:pedantic dependency and replace it with package:lints.
 
 ## 7.2.0