+++ scratch_space/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+- Drop package:pedantic dependency and replace it with package:lints.
+
 ## 1.0.0
 
 - Migrate to null-safety